flask-backend/requirements.txt
flask
python-dotenv
psycopg2
flask-cors
